requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pydantic==2.5.2
pydantic-settings==2.1.0 